New minor project milestone done. Will use this week to redesign most systems and refactor the graphics code and initialize graphics resources parallel.

+++ AEngine/AEngine/InputSystem.h
@@ -32,5 +32,5 @@ private:
 private:
     AEngine* m_pAEngine;
     bool m_movementBtnStates[4];
-    unsigned long m_movementScancodes[4];
+    unsigned short m_movementScancodes[4];
 };
